TI: HACK: remoteproc: k3-r5: Drop stale messages after suspend
It is observed that remoteproc firmware sends out a stale message after
responding to the suspend request from Linux. This causes a crash in
race conditions when we have already removed the rpmsg device after
suspend but have to process this stale message.

Unable to handle kernel NULL pointer dereference at virtual address
[  110.023229] CPU: 0 UID: 0 PID: 655 Comm: irq/662-mbox-ma Not tainted
[  110.031732] Hardware name: Texas Instruments J784S4 EVM (DT)
[  110.037374] pstate: 60000005 (nZCv daif -PAN -UAO -TCO -DIT -SSBS)
[  110.044317] pc : virtqueue_add_split+0x47c/0x644
[  110.048933] lr : virtqueue_add_split+0x3ac/0x644
[  110.053535] sp : ffff800085aebbd0
[  110.128023] Call trace:
[  110.130456]  virtqueue_add_split+0x47c/0x644
[  110.134713]  virtqueue_add_inbuf+0x48/0x54
[  110.138797]  rpmsg_recv_done+0x164/0x2e4
[  110.142709]  vring_interrupt+0x68/0xa4
[  110.146446]  rproc_vq_interrupt+0x2c/0x48
[  110.150444]  k3_r5_rproc_mbox_callback+0x110 [ti_k3_r5_remoteproc]
[  110.157134]  mbox_chan_received_data+0x1c/0x2c
[  110.161566]  mbox_interrupt+0xa0/0x17c [omap_mailbox]
[  110.166603]  irq_thread_fn+0x2c/0xa8
[  110.170168]  irq_thread+0x174/0x2fc
[  110.173644]  kthread+0x114/0x118
[  110.176863]  ret_from_fork+0x10/0x20

As a workaround, check if any message is received after suspend sequence
has started and return early in the mbox callback handler.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>