TI: HACK: remoteproc: k3-dsp: Don't send suspend msg for J7* devices
The firmware on C7x DSPs enter into a bad state after responding to
suspend message. As a workaround, do not send the suspend mailbox
message from Linux, assume it is ready for suspend and only send the
final shutdown message.

Introduce a device data flag 'fw_suspend_handshake' and send the suspend
message to rproc only if this flag is set. This flag is set to false for
all J7* devices in the driver data.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>